requirements.txt
alembic==1.15.1
appnope==0.1.4
asttokens==3.0.0
category-encoders==2.6.4
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.0
cramjam==2.9.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
fastparquet==2024.11.0
fonttools==4.56.0
fsspec==2025.3.0
geographiclib==2.0
geojson==2.5.0
geopy==2.4.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lightgbm==4.6.0
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
narwhals==1.29.1
nest-asyncio==1.6.0
numpy==2.0.2
optuna==4.2.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
plotly-express==0.4.1
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==25.1.0
pyarrow==19.0.1
Pygments==2.19.1
pyowm==3.3.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
scikit-learn==1.5.2
scikit-optimize==0.10.2
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.38
stack-data==0.6.3
statsmodels==0.14.4
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
xgboost==2.1.4
zipp==3.21.0
